Clean up: move BaseDevice out of basedevicemodel.h
